split config files into test/prod
