+++ shared/params/config.go
@@ -2,10 +2,7 @@
 package params
 
 import (
-	"sync"
 	"time"
-
-	"github.com/mohae/deepcopy"
 )
 
 // BeaconChainConfig contains constant configs for node to participate in beacon chain.
@@ -117,35 +114,4 @@ type BeaconChainConfig struct {
 
 	// Weak subjectivity values.
 	SafetyDecay uint64 // SafetyDecay is defined as the loss in the 1/3 consensus safety margin of the casper FFG mechanism.
-}
-
-var beaconConfig = MainnetConfig()
-var beaconConfigLock sync.RWMutex
-
-// BeaconConfig retrieves beacon chain config.
-func BeaconConfig() *BeaconChainConfig {
-	beaconConfigLock.RLock()
-	defer beaconConfigLock.RUnlock()
-	return beaconConfig
-}
-
-// OverrideBeaconConfig by replacing the config. The preferred pattern is to
-// call BeaconConfig(), change the specific parameters, and then call
-// OverrideBeaconConfig(c). Any subsequent calls to params.BeaconConfig() will
-// return this new configuration.
-func OverrideBeaconConfig(c *BeaconChainConfig) {
-	beaconConfigLock.Lock()
-	defer beaconConfigLock.Unlock()
-	beaconConfig = c
-}
-
-// Copy returns a copy of the config object.
-func (c *BeaconChainConfig) Copy() *BeaconChainConfig {
-	beaconConfigLock.RLock()
-	defer beaconConfigLock.RUnlock()
-	config, ok := deepcopy.Copy(*c).(BeaconChainConfig)
-	if !ok {
-		config = *beaconConfig
-	}
-	return &config
 }
