+++ shared/params/config_utils_prod.go
@@ -1,0 +1,34 @@
+// +build !test
+
+package params
+
+import (
+	"fmt"
+
+	"github.com/mohae/deepcopy"
+)
+
+var beaconConfig = MainnetConfig()
+
+// BeaconConfig retrieves beacon chain config.
+func BeaconConfig() *BeaconChainConfig {
+	fmt.Printf("Production level no locking!!!! ----------------------------->\n")
+	return beaconConfig
+}
+
+// OverrideBeaconConfig by replacing the config. The preferred pattern is to
+// call BeaconConfig(), change the specific parameters, and then call
+// OverrideBeaconConfig(c). Any subsequent calls to params.BeaconConfig() will
+// return this new configuration.
+func OverrideBeaconConfig(c *BeaconChainConfig) {
+	beaconConfig = c
+}
+
+// Copy returns a copy of the config object.
+func (c *BeaconChainConfig) Copy() *BeaconChainConfig {
+	config, ok := deepcopy.Copy(*c).(BeaconChainConfig)
+	if !ok {
+		config = *beaconConfig
+	}
+	return &config
+}
